+++ apps/typeform/frontend/src/Auth/TypeformOAuth.tsx
@@ -1,13 +1,13 @@
 import React, { useEffect, useState } from 'react';
 import { AppExtensionSDK } from '@contentful/app-sdk';
-import { Button } from '@contentful/forma-36-react-components';
-import { BASE_URL, CLIENT_ID } from '../constants';
+import { Button } from '@contentful/f36-components';
+import { BASE_URL, getClientId } from '../constants';
 
 interface Props {
   sdk?: AppExtensionSDK;
   expireSoon?: boolean;
   isFullWidth: boolean;
-  buttonType?: 'primary' | 'positive' | 'negative' | 'muted' | 'naked' | undefined;
+  buttonType?: 'primary' | 'positive' | 'negative' | 'secondary' | 'transparent' | undefined;
   setToken: (token: string) => void;
 }
 
@@ -57,7 +57,8 @@ export function TypeformOAuth({
   }, [oauthWindow, setToken]);
 
   const executeOauth = () => {
-    const url = `${BASE_URL}/oauth/authorize?&client_id=${CLIENT_ID}&redirect_uri=${encodeURIComponent(
+    const client_id = getClientId();
+    const url = `${BASE_URL}/oauth/authorize?&client_id=${client_id}&redirect_uri=${encodeURIComponent(
       `${window.location.origin}/callback`
     )}&scope=forms:read+workspaces:read`;
 
@@ -65,7 +66,7 @@ export function TypeformOAuth({
   };
 
   return (
-    <Button onClick={executeOauth} isFullWidth={isFullWidth} buttonType={buttonType} {...rest}>
+    <Button onClick={executeOauth} isFullWidth={isFullWidth} variant={buttonType} {...rest}>
       {expireSoon ? 'Reauthenticate with Typeform' : 'Sign in to Typeform'}
     </Button>
   );