+++ apps/typeform/frontend/src/constants.ts
@@ -2,6 +2,19 @@ const SDK_WINDOW_HEIGHT = 450;
 const BASE_URL = 'https://api.typeform.com';
 
 // TODO Move to env var
-const CLIENT_ID = 'HC3UDnoiaP1UCMqJ7kCAyTFdHrDt8nLtXx4BKRJxom2M';
+const TEST_CLIENT_ID = '7EugkySJjsYnSihDiYkGnMXrzbLBz7K8CFXS438Toqix';
+const PROD_CLIENT_ID = 'HC3UDnoiaP1UCMqJ7kCAyTFdHrDt8nLtXx4BKRJxom2M';
 
-export { SDK_WINDOW_HEIGHT, BASE_URL, CLIENT_ID };
+const getClientId = () => {
+  // if prod return prod client id
+  if (process.env.NODE_ENV === 'production') {
+    return PROD_CLIENT_ID;
+  } else if (process.env.NODE_ENV === 'staging') {
+    return TEST_CLIENT_ID;
+  } else {
+    console.error('Unknown environment: ', process.env.NODE_ENV);
+    return null;
+  }
+};
+
+export { SDK_WINDOW_HEIGHT, BASE_URL, getClientId };